requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-api-core==2.15.0
google-auth==2.27.0
aiohttp==3.9.1
asyncio==3.4.3
gunicorn==21.2.0 